fix(#342): apply fix pass changes
